Remove integrity check V1 and ledger dumps upload
CHANGELOG_BEGIN
CHANGELOG_END
