--- ledger/participant-state/kvutils/tools/integrity-check-v1/src/main/resources/logback.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-      <encoder>
-          <pattern>%level %logger{10}: %msg%n</pattern>
-      </encoder>
-  </appender>
-
-  <root level="INFO">
-      <appender-ref ref="STDOUT"/>
-  </root>
-</configuration>
